Make 2 component sources the default
This should have been done a while ago, since all of the testing
since June 2020 has been performed on two component sources
(since they preoved to be faster and more robust to model than
single component sources). So this commit just makes the default
the correct value to use.

+++ python/lsst/meas/extensions/scarlet/scarletDeblendTask.py
@@ -310,7 +310,7 @@ class ScarletDeblendConfig(pexConfig.Config):
         dtype=bool, default=True,
         doc="Whether or not to process isolated sources in the deblender")
     sourceModel = pexConfig.Field(
-        dtype=str, default="single",
+        dtype=str, default="double",
         doc=("How to determine which model to use for sources, from\n"
              "- 'single': use a single component for all sources\n"
              "- 'double': use a bulge disk model for all sources\n"
